Fix search query not being copied to clipboard on initiateSearchImmediately searches
Fixes #61

When initiating a search on page load, the search query was not being
copied to the clipboard as expected. This was due to using the wrong
query variable (`query` rather than `currentQuery`) being used in handleSearch, in SearchContext.

Also removed some leftover console.log statements in the SearchBar
component.